% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mvmcombC}
\alias{mvmcombC}
\title{Multiply applications of `vmcombC`.}
\usage{
mvmcombC(ns, ps)
}
\arguments{
\item{ns}{Names.}

\item{ps}{positions.}
}
\description{
Multiply applications of `vmcombC`.
}
