% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{par_distC}
\alias{par_distC}
\title{Parallel distance calculations.}
\usage{
par_distC(cols, mat)
}
\arguments{
\item{cols}{A range of column indexes.}

\item{mat}{A logical matrix.}
}
\description{
Parallel distance calculations.
}
