% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{to_lgldistC}
\alias{to_lgldistC}
\title{Extracts the lower triangular of a logical distance matrix.}
\usage{
to_lgldistC(m)
}
\arguments{
\item{m}{A logical matrix.}
}
\description{
Extracts the lower triangular of a logical distance matrix.
}
